Various additions and cleanups
Unified return types, so that confidence can be read by the user of a
prediction service.
Renamed classifiers and preditctors to include the algorithm name.
Added cross-vaidations for services (JSAT CV does not work yet).

+++ src/main/java/co/nz/leonhardt/bpe/logs/CaseLog.java
@@ -26,7 +26,7 @@ public class CaseLog {
 	private final Date timestamp;
 	
 	/* Amount requested */
-	private Double amountRequested; //TODO this is actually application-specific
+	private Long amountRequested; //TODO this is actually application-specific
 	
 	/**
 	 * Creates a new case log with the given uuid and timestamp.
@@ -57,11 +57,11 @@ public class CaseLog {
 		return timestamp;
 	}
 
-	public Double getAmountRequested() {
+	public Long getAmountRequested() {
 		return amountRequested;
 	}
 
-	public void setAmountRequested(Double amountRequested) {
+	public void setAmountRequested(Long amountRequested) {
 		this.amountRequested = amountRequested;
 	}
